+++ src/ot.js
@@ -1,0 +1,5 @@
+// This file is the central point where the OT types are imported.
+// Localized to one file so it's easy to change it in future.
+import OTJSON1Presence from 'sharedb-client-browser/dist/ot-json1-presence-umd.cjs';
+const { json1Presence, textUnicode } = OTJSON1Presence;
+export { json1Presence, textUnicode };
